First pass in cleaning up commerce slack views
