+++ lib/views/commerce_slack_view.ex
@@ -1,266 +1,12 @@
 # https://github.com/artsy/exchange/blob/master/app/events/order_event.rb#L1
 
 defmodule Aprb.Views.CommerceSlackView do
-  import Aprb.ViewHelper
-
+  alias Aprb.Views.{CommerceTransactionSlackView, CommerceOfferSlackView, CommerceOrderSlackView}
   def render(event, routing_key) do
-    case routing_key do
-      "transaction.created" -> failed_transaction_event(event)
-      "offer.submitted" -> offer_submitted(event)
-      "offer.pending_response" -> offer_pending_response(event)
-      _ -> order_event(event)
+    cond do
+      routing_key =~ "transaction." -> CommerceTransactionSlackView.render(event, routing_key)
+      routing_key =~ "offer." -> CommerceOfferSlackView.render(event, routing_key)
+      routing_key =~ "order." -> CommerceOrderSlackView.render(event, routing_key)
     end
-  end
-
-  defp failed_transaction_event(event) do
-    %{
-      text: ":alert: Failed Transaction",
-      attachments: [%{
-        fields: [
-          %{
-            title: "Failure Code",
-            value: event["properties"]["failure_code"],
-            short: true
-          },
-          %{
-            title: "Failure Message",
-            value: event["properties"]["failure_message"],
-            short: true
-          },
-          %{
-            title: "Transaction Type",
-            value: event["properties"]["transaction_type"],
-            short: true
-          },
-        ] ,
-        actions: [
-          %{
-            type: "button",
-            text: "Admin Link",
-            url: exchange_admin_link(event["properties"]["order"]["id"])
-          }
-        ]
-      }],
-      unfurl_links: true
-    }
-  end
-
-  defp offer_submitted(event) do
-    case Map.get(event["properties"], "in_response_to", nil) do
-      nil -> nil
-      _ -> counter_offer_view(event)
-    end
-  end
-
-  defp offer_pending_response(event) do
-    seller = fetch_info(event["properties"]["order"]["seller_id"], event["properties"]["order"]["seller_type"])
-    buyer = fetch_info(event["properties"]["order"]["buyer_id"], event["properties"]["order"]["buyer_type"])
-    %{
-      text: ":hourglass: Waiting Offer Response",
-      attachments:
-        line_item_attachments(event["properties"]["order"]["line_items"]) ++
-        [%{
-          fields: [
-            %{
-              title: "Offer Amount",
-              value: format_price(event["properties"]["amount_cents"] / 100),
-              short: true
-            },
-            %{
-              title: "By",
-              value: event["properties"]["from_participant"],
-              short: true
-            },
-            %{
-              title: "List Price",
-              value: format_price(event["properties"]["order"]["total_list_price_cents"] / 100),
-              short: true
-            },
-            %{
-              title: "Seller",
-              value: seller["name"],
-              short: true
-            },
-            %{
-              title: "Buyer",
-              value: cleanup_name(buyer["name"]),
-              short: true
-            }
-          ],
-          actions: [
-            %{
-              type: "button",
-              text: "Admin Link",
-              url: exchange_admin_link(event["properties"]["order"]["id"])
-            }
-          ]
-        }]
-    }
-  end
-
-  defp counter_offer_view(event) do
-    seller = fetch_info(event["properties"]["order"]["seller_id"], event["properties"]["order"]["seller_type"])
-    buyer = fetch_info(event["properties"]["order"]["buyer_id"], event["properties"]["order"]["buyer_type"])
-    %{
-      text: ":parrotsunnies: Counteroffer submitted",
-      attachments:
-        line_item_attachments(event["properties"]["order"]["line_items"]) ++
-        [%{
-          fields: [
-            %{
-              title: "Offer Amount",
-              value: format_price(event["properties"]["amount_cents"] / 100),
-              short: true
-            },
-            %{
-              title: "By",
-              value: event["properties"]["from_participant"],
-              short: true
-            },
-            %{
-              title: "Counter to",
-              value: format_price(event["properties"]["in_response_to"]["amount_cents"] / 100),
-              short: true
-            },
-            %{
-              title: "List Price",
-              value: format_price(event["properties"]["order"]["total_list_price_cents"] / 100),
-              short: true
-            },
-            %{
-              title: "Seller",
-              value: seller["name"],
-              short: true
-            },
-            %{
-              title: "Buyer",
-              value: cleanup_name(buyer["name"]),
-              short: true
-            }
-          ],
-          actions: [
-            %{
-              type: "button",
-              text: "Admin Link",
-              url: exchange_admin_link(event["properties"]["order"]["id"])
-            }
-          ]
-        }]
-    }
-  end
-
-  defp order_event(event) do
-    event
-      |> get_title
-      |> build_message(event)
-  end
-
-  defp get_title(event) do
-    case {event["verb"], event["properties"]["mode"]} do
-      {"submitted", "buy"} -> "🤞 Submitted"
-      {"submitted", "offer"} -> "🤞 Offer Submitted"
-      {"approved", _} -> ":yes: Approved"
-      {"canceled", _} ->
-        case event["properties"]["state_reason"] do
-          "seller_lapsed" -> ":zzz: Seller Lapsed"
-          "seller_rejected" -> ":soshocked: Seller Rejected"
-          "seller_rejected_offer_too_low" ->  ":soshocked: Seller Rejected - Offer too low"
-          "seller_rejected_shipping_unavailable" ->  ":soshocked: Seller Rejected - Shipping unavailable"
-          "seller_rejected_artwork_unavailable" ->  ":soshocked: Seller Rejected - Artwork unavailable"
-          "seller_rejected_other" ->  ":soshocked: Seller Rejected - other"
-          "buyer_rejected" ->  ":soshocked: Buyer Rejected"
-          "buyer_lapsed" ->  ":zzz: Buyer Lapsed"
-        end
-      {"refunded", _} -> ":sad-parrot: Refunded"
-      {"fulfilled", _} -> " :shipitmoveit: Fulfilled"
-      {"pending_approval", _} -> ":hourglass: Waiting Approval"
-      {"pending_fulfillment", _} -> ":hourglass: Waiting Shipping"
-      _ -> nil
-    end
-  end
-
-  defp build_message(nil, _event), do: nil
-  defp build_message(title, event) do
-    seller = fetch_info(event["properties"]["seller_id"], event["properties"]["seller_type"])
-    buyer = fetch_info(event["properties"]["buyer_id"], event["properties"]["buyer_type"])
-    %{
-      text: "#{title} #{artworks_links_from_line_items(event["properties"]["line_items"])}",
-      attachments: order_attachments(event["properties"], event["object"]["id"], seller, buyer),
-      unfurl_links: true
-    }
-  end
-
-  defp order_attachments(order_properties, order_id, seller, buyer) do
-    fields = order_attachment_fields(order_properties, seller, buyer)
-      |> append_admin(seller["admin"])
-      |> append_offer_fields(order_properties["mode"], order_properties)
-    [
-      %{
-        color: "#6E1FFF",
-        author_name: order_properties["code"],
-        author_link: exchange_admin_link(order_id),
-        title: seller["name"],
-        title_link: admin_partners_link("partners/#{seller["id"]}"),
-        fields: fields
-      }
-    ] ++ line_item_attachments(order_properties["line_items"])
-  end
-
-  defp append_admin(attachments, nil), do: attachments
-  defp append_admin(attachments, admin), do: attachments ++ [%{ title: "Admin", value: admin["name"], short: true}]
-
-  defp append_offer_fields(attachments, "offer", properties), do: attachments ++ [%{title: "List Price", value: format_price(properties["total_list_price_cents"] / 100), short: true}]
-  defp append_offer_fields(attachments, _, _), do: attachments
-
-  defp order_attachment_fields(order_properties, seller, buyer) do
-    [
-      %{
-        title: "Purchase Method",
-        value: order_properties["mode"],
-        short: true
-      },
-      %{
-        title: "Buyer",
-        value: cleanup_name(buyer["name"]),
-        short: true
-      },
-      %{
-        title: "Total Amount",
-        value: format_price(order_properties["items_total_cents"] / 100),
-        short: true
-      }
-    ]
-  end
-
-  defp fetch_info(id, "user"), do: Gravity.get!("/users/#{id}").body
-  defp fetch_info(id, _), do: Gravity.get!("/v1/partner/#{id}").body
-
-  defp artworks_links_from_line_items(line_items) do
-    line_items
-      |> Enum.map(fn(li) -> artwork_link(li["artwork_id"]) end)
-      |> Enum.map(fn(artwork_link) -> "<#{artwork_link}| >" end)
-  end
-
-  defp line_item_attachments(line_items) do
-    line_items
-      |> Enum.map(&line_item_attachment(&1))
-  end
-
-  defp line_item_attachment(line_item) do
-    artwork = Gravity.get!("/v1/artwork/#{line_item["artwork_id"]}").body
-    %{
-      fields: [
-        %{
-          title: "Available Purchase Modes",
-          value: cond do
-                    artwork["ecommerce"] && artwork["offer"] -> "BNMO"
-                    artwork["ecommerce"] -> "BN"
-                    artwork["offer"] -> "MO"
-                    true -> "!?"
-                 end,
-          short: true
-        }
-      ]
-    }
   end
 end
